+++ packages/impromat-api/src/graphql/services/element.service.ts
@@ -6,7 +6,7 @@ import {
 import { PrismaService } from './prisma.service';
 
 import { accessibleBy } from '@casl/prisma';
-import { Prisma } from '@prisma/client';
+import { ElementVisibility, Prisma } from '@prisma/client';
 import { ElementsOrderByInput } from 'src/dtos/inputs/elements-query-input';
 import { ElementsFilterInput } from 'test/graphql-client/graphql';
 import { defineAbilityFor } from '../abilities';
@@ -101,13 +101,46 @@ export class ElementService {
     updateElementInput: UpdateElementInput,
   ) {
     const existing = await this.prismaService.element.findFirstOrThrow({
-      where: { id: updateElementInput.id, ownerId: userRequestId },
+      where: {
+        AND: [
+          { id: updateElementInput.id },
+          {
+            OR: [
+              { ownerId: userRequestId },
+              { visibility: ElementVisibility.PUBLIC },
+            ],
+          },
+        ],
+      },
+      include: {
+        tags: true,
+      },
     });
     if (!existing) throw new Error('Not existing or not owner.');
 
-    return this.prismaService.element.update({
-      where: { id: updateElementInput.id },
-      data: updateElementInput,
+    return await this.prismaService.$transaction(async (tx) => {
+      await tx.element.create({
+        data: {
+          ...existing,
+          ...{
+            snapshotParentId: existing.id,
+            id: undefined,
+            updatedAt: undefined,
+            createdAt: undefined,
+            improbibIdentifier: undefined,
+            tags: {
+              connect: existing.tags.map((existingTag) => ({
+                id: existingTag.id,
+              })),
+            },
+          },
+        },
+      });
+
+      return tx.element.update({
+        where: { id: updateElementInput.id },
+        data: updateElementInput,
+      });
     });
   }
 }